python/datacatalog/requirements.txt
google-cloud-datacatalog==3.19.0
google-cloud-bigquery_datapolicies==0.6.6
google-cloud-bigquery==3.20.1